sdcard_image-sunxi: add configurable behaviour to prevent U-Boot reflash
When using the full U-Boot + SPL binary on 64bit boards, flashing again
the U-Boot binary invalidates the image, making the board stuck on
the SPL. This can now be disabled by setting a variable.

+++ classes/sdcard_image-sunxi.bbclass
@@ -121,7 +121,7 @@ IMAGE_CMD:sunxi-sdimg () {
 
 # write uboot.itb for arm64 boards
 IMAGE_CMD_sunxi-sdimg:append:sun50i () {
-	if [ -e "${DEPLOY_DIR_IMAGE}/${UBOOT_BINARY}" ]
+	if [ -e "${DEPLOY_DIR_IMAGE}/${UBOOT_BINARY}" -a "${USE_SPLIT_UBOOT_SPL}" = "1"  ]
 	then
 		dd if=${DEPLOY_DIR_IMAGE}/${UBOOT_BINARY} of=${SDIMG} bs=1024 seek=40 conv=notrunc
 	fi
